Adding scripts uri 1038 ruby

+++ ruby/app/code/1038.rb
@@ -26,7 +26,7 @@ if __FILE__ == $0
     valores = ((STDIN.gets).strip).split(" ")
     array = $table
     if valores.size == 2
-        first = (($table[((valores[0]).to_i) - 1]).last()).to_i
+        first = (($table[((valores[0]).to_i) - 1]).last()).to_f
         count = (valores[1]).to_i
         puts "#{$prices}"
         puts "#{first}"
